requirements.txt
torch==2.5.1
torchvision==0.20.1
accelerate==0.33.0
ema-pytorch==0.5.3
denoising-diffusion-pytorch==2.1.1
tensorboard==2.16.2
tensorboardX==2.6.2.2
triton==3.1.0
numpy==1.26.4
scipy==1.14.0
scikit-learn==1.5.2
pandas==2.2.2
h5py==3.12.1
zarr==3.0.0
numcodecs==0.14.1
einops==0.8.0
neuraloperator==1.0.2
torch_harmonics==0.7.3
tensorly==0.9.0
tensorly-torch==0.5.0
matplotlib==3.9.2
seaborn==0.13.2
imageio==2.37.0
svgwrite==1.4.3
pillow==10.3.0
tqdm==4.65.0
pydantic==2.10.4
wandb==0.19.4
PyYAML==6.0.1
ml-collections==1.1.0
fsspec==2024.5.0
filelock==3.14.0
safetensors==0.4.4
huggingface-hub==0.24.5
networkx==3.3
sympy==1.13.1
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cusolver-cu12==11.6.1.9